requirements.txt.txt
numpy==1.19.2
Flask==1.1.2
tensorflow_hub==0.11.0
tensorflow-cpu==2.5.0
gunicorn==20.0.4
Jinja2==2.11.2
Werkzeug==1.0.1
rq==1.8.0
redis==3.5.3